distribute/0.0.2/build_shell/teamcat/requirements.txt
django==2.0
mysqlclient
pillow
qrcode
httplib2
django-websocket-redis
uwsgi
pycrypto
coreapi
markdown
django-filter
django-crispy-forms
django-guardian
djangorestframework
django-url-filter
django-cors-headers
django-oauth-toolkit
mongoengine
blinker
pymongo
django-rest-framework-mongoengine
filetype
pytz
xlwt